Fix Sheets not showing '0' when Formatting is Automatic
If your cell formatting in Sheets is Automatic (which it is by default), uploading a record with an Integer property with value '0' will result in the cell remaining blank (but having the 0 value if you click on the cell).

This fix changes the cell formatting to '0', which results in the value '0' being visible.